refactor: move observability.go to comp/api to avoid adding import in all binaries
